update engine
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@1818 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/engine/localevent.cpp
@@ -80,7 +80,7 @@ void LocalEvent::SetTapMode(bool f)
     f ? SetModes(TAP_MODE) : ResetModes(TAP_MODE);
 }
 
-void LocalEvent::SetTapDelayForRightClickEmulation(double d)
+void LocalEvent::SetTapDelayForRightClickEmulation(u32 d)
 {
     clock_delay = d < 200 ? TAP_DELAY_EMULATE : d;
 }
